Update view and viewModel names, add state management for TaskListWithStateViewController
