+++ SimplyDone/SimplyDone/TaskDetails/TaskDetails.swift
@@ -2,143 +2,84 @@
 //  TaskDetails.swift
 //  SimplyDone
 //
-//  Created by Ramazan Abdullayev on 01/03/2025.
+//  Created by Ramazan Abdullayev on 16/03/2025.
 //
 
 import SwiftUI
+import RealmSwift
 
-struct TaskDetails: View {
+@Observable
+class TaskDetails {
 
-    private(set) var viewModel: TaskDetailsViewModel
+    let taskModel: TaskModel
+    let router: TaskDetailsRouter
+    let realm = try? Realm()
 
-    var body: some View {
-        VStack(alignment: .leading, spacing: 20) {
+    init(taskModel: TaskModel, router: TaskDetailsRouter) {
+        self.taskModel = taskModel
+        self.router = router
+    }
 
-            HStack(alignment: .center) {
-                Image(systemName: viewModel.taskModel.status == TaskStatus.completed.rawValue ? "checkmark.square.fill": "square")
-                    .imageScale(.large)
-                    .frame(width: 30, height: 30)
-                    .foregroundColor(Color(UIColor.darkGray))
-                    .onTapGesture {
-                        viewModel.taskModel.status.toggleStatus()
-                        viewModel.updateRealmTaskStatus()
-                    }
+    func deleteRealmSubtask(for index: Int) {
+        guard let task = realm?.objects(Task.self).filter({ [weak self] in
+            $0.id == self?.taskModel.id
+        }).first else { return }
 
-                Text(viewModel.taskModel.title)
-                    .font(.title)
-                    .fontWeight(.bold)
+        let subtaskModel = taskModel.subtasks[index]
+        let subtasksToDelete = Array(task.subtasks).filter({ $0.id == subtaskModel.id })
+        do {
+            try realm?.write {
+                realm?.delete(subtasksToDelete)
             }
-            .padding(.top, 20)
-
-            Text(viewModel.taskModel.taskDescription)
-                .font(.body)
-                .foregroundColor(.secondary)
-                .padding()
-
-            HStack {
-                Image(systemName: viewModel.taskModel.dueDate == nil ? "calendar.badge.exclamationmark" : "calendar")
-                Text("\(viewModel.taskModel.dueDate?.formatted(date: .abbreviated, time: .omitted) ?? "No Due Date")")
-                    .font(.subheadline)
-                    .foregroundColor(.secondary)
-            }
-            .padding(.top, 10)
-
-            Divider()
-
-            subtasksList
-
-            Spacer()
-        }
-        .padding()
-        .navigationTitle("Task Details")
-        .toolbar {
-            ToolbarItem(placement: .navigationBarTrailing) {
-                Image(systemName: "square.and.pencil")
-                    .imageScale(.large)
-                    .foregroundColor(.black)
-                    .onTapGesture { 
-                        viewModel.navigateToEditTask()
-                    }
-            }
+        } catch {
+            // Handle the error here
         }
     }
 
-    private var subtasksList: some View {
-        Group {
-            HStack(alignment: .center) {
-                Text("Subtasks")
-                    .font(.title2)
-                    .fontWeight(.bold)
-                
-                Spacer()
+    func updateRealmTaskStatus() {
+        guard let task = realm?.objects(Task.self).filter({ [weak self] in
+            $0.id == self?.taskModel.id
+        }).first else { return }
 
-                Image(systemName: "plus")
-                    .imageScale(.large)
-                    .frame(width: 30, height: 30)
-                    .foregroundColor(.black)
-                    .onTapGesture {
-                        viewModel.navigateToAddNewSubtask()
-                    }
+        do {
+            try realm?.write {
+                task.status = taskModel.status
             }
-            .padding(.horizontal, 10)
-
-            if !viewModel.taskModel.subtasks.isEmpty {
-                VStack(alignment: .leading, spacing: 10) {
-                    List {
-                        ForEach(viewModel.taskModel.subtasks, id: \.self) { subtaskModel in
-                            getSubtaskView(subtaskModel: subtaskModel)
-                        }
-                        .onDelete { indexSet in
-                            guard let firstIndex = indexSet.first else { return }
-                            viewModel.deleteRealmSubtask(for: firstIndex)
-                        }
-                    }
-                    .listStyle(.plain)
-                }
-            } else {
-                VStack(alignment: .center) {
-                    Image(systemName: "tray.fill")
-                        .resizable()
-                        .scaledToFit()
-                        .frame(width: 50, height: 50)
-                        .foregroundColor(.gray)
-                    
-                    Text("No subtasks")
-                        .font(.headline)
-                        .foregroundColor(.gray)
-                        .padding(.top, 5)
-                }
-                .frame(maxWidth: .infinity, minHeight: 150)
-            }
+        } catch {
+            // Handle the error here
         }
     }
 
-    private func getSubtaskView(subtaskModel: SubtaskModel) -> some View {
-        HStack {
-            Image(systemName: subtaskModel.status == TaskStatus.completed.rawValue ? "checkmark.square.fill" : "square")
-                .foregroundColor(.gray)
-                .padding(.trailing, 10)
-                .onTapGesture {
-                    if let subtaskIndex = viewModel.taskModel.subtasks.firstIndex(where: { $0.id == subtaskModel.id }) {
-                        self.viewModel.taskModel.subtasks[subtaskIndex].status.toggleStatus()
-                        self.viewModel.updateRealmSubtaskStatus(for: subtaskIndex)
-                    }
-                }
+    func updateRealmSubtaskStatus(for index: Int) {
+        guard let task = realm?.objects(Task.self).filter({ [weak self] in
+            $0.id == self?.taskModel.id
+        }).first else { return }
 
-            Text(subtaskModel.title)
-                .font(.body)
-                .strikethrough(subtaskModel.status == TaskStatus.completed.rawValue, color: .gray)
-        }
-        .padding(.vertical, 5)
-        .onTapGesture {
-            viewModel.navigateToSubtaskDetails(for: subtaskModel)
+        let subtaskModel = taskModel.subtasks[index]
+        guard let subtask = Array(task.subtasks).filter({
+            $0.id == subtaskModel.id
+        }).first else { return }
+
+        do {
+            try realm?.write {
+                subtask.status = subtaskModel.status
+            }
+        } catch {
+            // Handle the error here
         }
     }
-}
 
-#Preview {
-    TaskDetails(viewModel: TaskDetailsViewModel(
-        taskModel: TaskModel.generatedTaskModels.first!,
-        router: .init()
-    ))
+    // MARK: - Navigation
+
+    public func navigateToEditTask() {
+        router.navigateToEditTask(for: taskModel)
+    }
+
+    public func navigateToAddNewSubtask() {
+        router.navigateToAddNewSubtask(for: taskModel)
+    }
+
+    public func navigateToSubtaskDetails(for subtask: SubtaskModel) {
+        router.navigateToSubtaskDetails(for: taskModel, and: subtask)
+    }
 }
